refactor: extract shared Sugar order-row methods into SugarCRMOrderRowsTrait
RepairSugarOrderPurchasePrices extended ImportOrdersFromSugarCRM only to
access shared utility methods, but inherited handle() references --date-to/
--date-from options not declared in the child command signature, triggering
a Larastan false positive.

Extract the 12 shared methods (fetchOrderRows, resolveProductForSugarRow,
productsByNameForSugarRows, partnerProductsBy*, normalizeProductName,
orderItemMainPurchasePayloadFromSugarRow, buildPurchasePayloadFromSugarAmounts,
sugarMoneyAmount, sugarOrderRowCstm*) into a new SugarCRMOrderRowsTrait.
Both commands now extend AbstractSugarCRMImport and use the trait, removing
the inheritance relationship that caused the PHPStan error.
